+++ client/src/interfaces/Forms.ts
@@ -27,4 +27,5 @@ export interface OtherSettingsForm {
   kubernetesApps: boolean;
   unpinStoppedApps: boolean;
   useAmericanDate: boolean;
+  greetingsSchema: string;
 }