+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### v1.7.4 (TBA)
+- [WIP] Added option to set custom greetings and date ([#103](https://github.com/pawelmalak/flame/issues/103))
 - Added iOS "Add to homescreen" icon ([#131](https://github.com/pawelmalak/flame/issues/131))
+- Added 3 new themes
 
 ### v1.7.3 (2021-10-28)
 - Fixed bug with custom CSS not updating